SamTranscriptomeParser & DefinedRegionDifferentialSeq * Replaced the IH tag with NH tag for reporting the number of alignments present per read to be compatible with DEXSeq's HTSeq app.
SamReadDepthSubSampler
* New app for reducing extreme read depths in amplicon based sequencing datasets

More mods to the SpliceJunction annotator for zscore estimations

+++ trunk/Source/edu/utah/seq/base/NMer.java
@@ -192,7 +192,7 @@ public class NMer {
 			return sb.toString();
 		}
 
-		public long getNumberObservations(){
+		public double getNumberObservations(){
 			return sdA[0].getNumberObservations();
 		}
